+++ consensus/src/threshold_simplex/actors/resolver/ingress.rs
@@ -25,7 +25,7 @@ pub struct Mailbox<D: Digest> {
 }
 
 impl<D: Digest> Mailbox<D> {
-    pub(crate) fn new(sender: mpsc::Sender<Message<D>>) -> Self {
+    pub fn new(sender: mpsc::Sender<Message<D>>) -> Self {
         Self { sender }
     }
