+++ consensus/src/threshold_simplex/actors/voter/ingress.rs
@@ -24,7 +24,7 @@ pub struct Mailbox<D: Digest> {
 }
 
 impl<D: Digest> Mailbox<D> {
-    pub(super) fn new(sender: mpsc::Sender<Message<D>>) -> Self {
+    pub fn new(sender: mpsc::Sender<Message<D>>) -> Self {
         Self { sender }
     }
